HUCO-86 | [feat] card.cs and customcard.cs merge into card.cs
