--- HandsUp/Assets/CardViewScene/Scripts/Card/CustomCard.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 26a3b7b3b3ed30f49ae274a7be301e74
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
